--- app/components/scihist_constraints_component.rb
@@ -1,33 +1,0 @@
-# This overrides Blacklight::ConstraintsComponent ONLY for the purpose of over-riding the
-# default value for the initializer `query_constraint_component` keyword argument --
-#
-# So we can set our own custom ScihistQueryConstraintComponent, which is really what
-# we wanted to do.
-#
-# Overridden at https://github.com/projectblacklight/blacklight/blame/13a8122fc6495e52acabc33875b80b51613d8351/app/components/blacklight/constraints_component.rb
-class ScihistConstraintsComponent < Blacklight::ConstraintsComponent
-
-  # override just to change default value of `query_constraint_component`
-  def initialize(*, search_state:, query_constraint_component: ScihistQueryConstraintComponent, **)
-    @search_state = search_state
-    super
-  end
-
-  # Override from Blacklight::ConstraintsComponent to:
-  #
-  # * Show even if there is a blank query, since we want to show the form
-  #   so you can search within facet results.
-  #
-  # * This also lets us make sure we're controlling the params passed to
-  #   our custom QueryComponent, since that is local code that is not updated
-  #   by Blacklight!
-  #
-  # I don't entirely understand the original implementation where it adds
-  # on `+ helpers.render(@facet_constraint_component`, we're leaving that out... seems
-  # to be fine?  Sorry.
-  def query_constraints
-    helpers.render @query_constraint_component.new(
-      search_state: @search_state
-    )
-  end
-end
